Move HasPosition outside of CanPlay
